Fix payment migration user foreign key type

+++ database/migrations/2026_08_18_160100_create_server_purchase_payments_table.php
@@ -9,7 +9,8 @@ return new class () extends Migration {
     {
         Schema::create('server_purchase_payments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            // Pterodactyl's users table uses an unsigned INT primary key.
+            $table->unsignedInteger('user_id');
             $table->unsignedBigInteger('plan_id')->nullable();
             $table->string('reference', 191)->unique();
             $table->string('gateway', 32);
